Rename back to myq
