+++ custom_components/myq/binary_sensor.py
