+++ custom_components/myq/translations/ca.json
