+++ custom_components/myq/translations/cs.json
