+++ custom_components/myq/translations/da.json
