+++ custom_components/myq/translations/de.json
