+++ custom_components/myq/translations/en.json
