+++ custom_components/myq/translations/es-419.json
