+++ custom_components/myq/translations/es.json
