+++ custom_components/myq/translations/et.json
