+++ custom_components/myq/translations/fr.json
