+++ custom_components/myq/translations/hu.json
