+++ custom_components/myq/translations/ko.json
