+++ custom_components/myq/translations/lb.json
