+++ custom_components/myq/translations/nb.json
