+++ custom_components/myq/translations/no.json
