+++ custom_components/myq/translations/pt-BR.json
