+++ custom_components/myq/translations/pt.json
