+++ custom_components/myq/translations/sl.json
