+++ custom_components/myq/translations/sv.json
