+++ custom_components/myq/translations/zh-Hans.json
