+++ custom_components/myq/translations/zh-Hant.json
